--- a/src/KOKKOS/sna_kokkos.h
+++ b/src/KOKKOS/sna_kokkos.h
@@ -25,45 +25,78 @@
 
 namespace LAMMPS_NS {
 
-template<class DeviceType>
+template<typename real_, int vector_length_>
+struct WignerWrapper {
+  using real = real_;
+  using complex = SNAComplex<real>;
+  static constexpr int vector_length = vector_length_;
+
+  const int offset; // my offset into the vector (0, ..., vector_length - 1)
+  real* buffer; // buffer of real numbers
+
+  KOKKOS_INLINE_FUNCTION
+  WignerWrapper(complex* buffer_, const int offset_)
+   : offset(offset_), buffer(reinterpret_cast<real*>(buffer_))
+  { ; }
+
+  KOKKOS_INLINE_FUNCTION
+  complex get(const int& ma) {
+    return complex(buffer[offset + 2 * vector_length * ma], buffer[offset + vector_length + 2 * vector_length * ma]);
+  }
+
+  KOKKOS_INLINE_FUNCTION
+  void set(const int& ma, const complex& store) {
+    buffer[offset + 2 * vector_length * ma] = store.re;
+    buffer[offset + vector_length + 2 * vector_length * ma] = store.im;
+  }
+};
+
+struct alignas(8) FullHalfMapper {
+  int idxu_half;
+  int flip_sign; // 0 -> isn't flipped, 1 -> conj, -1 -> -conj
+};
+
+template<class DeviceType, typename real_, int vector_length_>
 class SNAKokkos {
 
 public:
+  using real = real_;
+  using complex = SNAComplex<real>;
+  static constexpr int vector_length = vector_length_;
+
   typedef Kokkos::View<int*, DeviceType> t_sna_1i;
-  typedef Kokkos::View<double*, DeviceType> t_sna_1d;
-  typedef Kokkos::View<double*, typename KKDevice<DeviceType>::value, Kokkos::MemoryTraits<Kokkos::Atomic> > t_sna_1d_atomic;
+  typedef Kokkos::View<real*, DeviceType> t_sna_1d;
+  typedef Kokkos::View<real*, typename KKDevice<DeviceType>::value, Kokkos::MemoryTraits<Kokkos::Atomic> > t_sna_1d_atomic;
   typedef Kokkos::View<int**, DeviceType> t_sna_2i;
-  typedef Kokkos::View<double**, DeviceType> t_sna_2d;
-  typedef Kokkos::View<double**, Kokkos::LayoutLeft, DeviceType> t_sna_2d_ll;
-  typedef Kokkos::View<double***, DeviceType> t_sna_3d;
-  typedef Kokkos::View<double***, Kokkos::LayoutLeft, DeviceType> t_sna_3d_ll;
-  typedef Kokkos::View<double***[3], DeviceType> t_sna_4d;
-  typedef Kokkos::View<double****, Kokkos::LayoutLeft, DeviceType> t_sna_4d_ll;
-  typedef Kokkos::View<double**[3], DeviceType> t_sna_3d3;
-  typedef Kokkos::View<double*****, DeviceType> t_sna_5d;
+  typedef Kokkos::View<real**, DeviceType> t_sna_2d;
+  typedef Kokkos::View<real**, Kokkos::LayoutLeft, DeviceType> t_sna_2d_ll;
+  typedef Kokkos::View<real***, DeviceType> t_sna_3d;
+  typedef Kokkos::View<real***, Kokkos::LayoutLeft, DeviceType> t_sna_3d_ll;
+  typedef Kokkos::View<real***[3], DeviceType> t_sna_4d;
+  typedef Kokkos::View<real****, Kokkos::LayoutLeft, DeviceType> t_sna_4d_ll;
+  typedef Kokkos::View<real**[3], DeviceType> t_sna_3d3;
+  typedef Kokkos::View<real*****, DeviceType> t_sna_5d;
 
-  typedef Kokkos::View<SNAcomplex*, DeviceType> t_sna_1c;
-  typedef Kokkos::View<SNAcomplex*, typename KKDevice<DeviceType>::value, Kokkos::MemoryTraits<Kokkos::Atomic> > t_sna_1c_atomic;
-  typedef Kokkos::View<SNAcomplex**, DeviceType> t_sna_2c;
-  typedef Kokkos::View<SNAcomplex**, Kokkos::LayoutLeft, DeviceType> t_sna_2c_ll;
-  typedef Kokkos::View<SNAcomplex**, Kokkos::LayoutRight, DeviceType> t_sna_2c_lr;
-  typedef Kokkos::View<SNAcomplex***, DeviceType> t_sna_3c;
-  typedef Kokkos::View<SNAcomplex***, Kokkos::LayoutLeft, DeviceType> t_sna_3c_ll;
-  typedef Kokkos::View<SNAcomplex***[3], DeviceType> t_sna_4c;
-  typedef Kokkos::View<SNAcomplex***[3], Kokkos::LayoutLeft, DeviceType> t_sna_4c3_ll;
-  typedef Kokkos::View<SNAcomplex****, Kokkos::LayoutLeft, DeviceType> t_sna_4c_ll;
-  typedef Kokkos::View<SNAcomplex**[3], DeviceType> t_sna_3c3;
-  typedef Kokkos::View<SNAcomplex*****, DeviceType> t_sna_5c;
-
-  typedef Kokkos::View<CayleyKleinPack**, DeviceType> t_sna_2ckp;
+  typedef Kokkos::View<complex*, DeviceType> t_sna_1c;
+  typedef Kokkos::View<complex*, typename KKDevice<DeviceType>::value, Kokkos::MemoryTraits<Kokkos::Atomic> > t_sna_1c_atomic;
+  typedef Kokkos::View<complex**, DeviceType> t_sna_2c;
+  typedef Kokkos::View<complex**, Kokkos::LayoutLeft, DeviceType> t_sna_2c_ll;
+  typedef Kokkos::View<complex**, Kokkos::LayoutRight, DeviceType> t_sna_2c_lr;
+  typedef Kokkos::View<complex***, DeviceType> t_sna_3c;
+  typedef Kokkos::View<complex***, Kokkos::LayoutLeft, DeviceType> t_sna_3c_ll;
+  typedef Kokkos::View<complex***[3], DeviceType> t_sna_4c;
+  typedef Kokkos::View<complex***[3], Kokkos::LayoutLeft, DeviceType> t_sna_4c3_ll;
+  typedef Kokkos::View<complex****, Kokkos::LayoutLeft, DeviceType> t_sna_4c_ll;
+  typedef Kokkos::View<complex**[3], DeviceType> t_sna_3c3;
+  typedef Kokkos::View<complex*****, DeviceType> t_sna_5c;
 
 inline
   SNAKokkos() {};
   KOKKOS_INLINE_FUNCTION
-  SNAKokkos(const SNAKokkos<DeviceType>& sna, const typename Kokkos::TeamPolicy<DeviceType>::member_type& team);
+  SNAKokkos(const SNAKokkos<DeviceType,real,vector_length>& sna, const typename Kokkos::TeamPolicy<DeviceType>::member_type& team);
 
 inline
-  SNAKokkos(double, int, double, int, int, int, int, int, int);
+  SNAKokkos(real, int, real, int, int, int, int, int, int);
 
   KOKKOS_INLINE_FUNCTION
   ~SNAKokkos();
@@ -81,17 +114,16 @@ inline
 
   // functions for bispectrum coefficients, GPU only
   KOKKOS_INLINE_FUNCTION
-  void compute_cayley_klein(const int&, const int&, const double&, const double&,
-                            const double&, const double&, const double&);
+  void compute_cayley_klein(const int&, const int&, const int&);
   KOKKOS_INLINE_FUNCTION
-  void pre_ui(const typename Kokkos::TeamPolicy<DeviceType>::member_type& team,const int&,const int&); // ForceSNAP
+  void pre_ui(const int&, const int&, const int&, const int&); // ForceSNAP
   KOKKOS_INLINE_FUNCTION
-  void compute_ui(const typename Kokkos::TeamPolicy<DeviceType>::member_type& team, const int, const int); // ForceSNAP
+  void compute_ui(const typename Kokkos::TeamPolicy<DeviceType>::member_type& team, const int, const int, const int, const int); // ForceSNAP
   KOKKOS_INLINE_FUNCTION
   void compute_zi(const int&, const int&, const int&);    // ForceSNAP
   KOKKOS_INLINE_FUNCTION
   void compute_yi(int,int,int,
-   const Kokkos::View<F_FLOAT***, Kokkos::LayoutLeft, DeviceType> &beta_pack); // ForceSNAP
+   const Kokkos::View<real***, Kokkos::LayoutLeft, DeviceType> &beta_pack); // ForceSNAP
   KOKKOS_INLINE_FUNCTION
   void compute_bi(const int&, const int&, const int&);    // ForceSNAP
 
@@ -104,34 +136,33 @@ inline
   void compute_zi_cpu(const int&);    // ForceSNAP
   KOKKOS_INLINE_FUNCTION
   void compute_yi_cpu(int,
-   const Kokkos::View<F_FLOAT**, DeviceType> &beta); // ForceSNAP
+   const Kokkos::View<real**, DeviceType> &beta); // ForceSNAP
     KOKKOS_INLINE_FUNCTION
   void compute_bi_cpu(const typename Kokkos::TeamPolicy<DeviceType>::member_type& team, int);    // ForceSNAP
 
   // functions for derivatives, GPU only
   KOKKOS_INLINE_FUNCTION
-  void compute_fused_deidrj(const typename Kokkos::TeamPolicy<DeviceType>::member_type& team, const int, const int); //ForceSNAP
+  void compute_fused_deidrj(const typename Kokkos::TeamPolicy<DeviceType>::member_type& team, const int, const int, const int, const int); //ForceSNAP
 
   // functions for derivatives, CPU only
   KOKKOS_INLINE_FUNCTION
   void compute_duidrj_cpu(const typename Kokkos::TeamPolicy<DeviceType>::member_type& team, int, int); //ForceSNAP
   KOKKOS_INLINE_FUNCTION
   void compute_deidrj_cpu(const typename Kokkos::TeamPolicy<DeviceType>::member_type& team, int, int); // ForceSNAP
-  KOKKOS_INLINE_FUNCTION
-  double compute_sfac(double, double); // add_uarraytot, compute_duarray
-  KOKKOS_INLINE_FUNCTION
-  double compute_dsfac(double, double); // compute_duarray
-  KOKKOS_INLINE_FUNCTION
-  void compute_s_dsfac(const double, const double, double&, double&); // compute_cayley_klein
 
-  // efficient complex FMA
-  // efficient caxpy (i.e., y += a x)
-  static KOKKOS_FORCEINLINE_FUNCTION
-  void caxpy(const SNAcomplex& a, const SNAcomplex& x, SNAcomplex& y);
+  KOKKOS_INLINE_FUNCTION
+  real compute_sfac(real, real); // add_uarraytot, compute_duarray
 
-  // efficient complex FMA, conjugate of scalar
+  KOKKOS_INLINE_FUNCTION
+  real compute_dsfac(real, real); // compute_duarray
+
+  KOKKOS_INLINE_FUNCTION
+  void compute_s_dsfac(const real, const real, real&, real&); // compute_cayley_klein
+
   static KOKKOS_FORCEINLINE_FUNCTION
-  void caconjxpy(const SNAcomplex& a, const SNAcomplex& x, SNAcomplex& y);
+  void sincos_wrapper(double x, double* sin_, double *cos_) { sincos(x, sin_, cos_); }
+  static KOKKOS_FORCEINLINE_FUNCTION
+  void sincos_wrapper(float x, float* sin_, float *cos_) { sincosf(x, sin_, cos_); }
 
   // Set the direction for split ComputeDuidrj
   KOKKOS_INLINE_FUNCTION
@@ -145,10 +176,6 @@ inline
 #endif
 
   //per sna class instance for OMP use
-
-  // Alternative to rij, wj, rcutij...
-  // just calculate everything up front
-  t_sna_2ckp cayleyklein;
 
   // Per InFlight Particle
   t_sna_3d rij;
@@ -175,8 +202,14 @@ inline
   t_sna_4c3_ll dulist;
 
   // Modified structures for GPU backend
-  t_sna_3d_ll ulisttot_re; // split real,
-  t_sna_3d_ll ulisttot_im; // imag
+  t_sna_3c_ll a_pack; // Cayley-Klein `a`
+  t_sna_3c_ll b_pack; // `b`
+  t_sna_4c_ll da_pack; // `da`
+  t_sna_4c_ll db_pack; // `db`
+  t_sna_4d_ll sfac_pack; // sfac, dsfac_{x,y,z}
+
+  t_sna_4d_ll ulisttot_re_pack; // split real,
+  t_sna_4d_ll ulisttot_im_pack; // imag, AoSoA, flattened
   t_sna_4c_ll ulisttot_pack; // AoSoA layout
   t_sna_4c_ll zlist_pack; // AoSoA layout
   t_sna_4d_ll blist_pack;
@@ -191,7 +224,7 @@ inline
   int ntriples;
 
 private:
-  double rmin0, rfac0;
+  real rmin0, rfac0;
 
   //use indexlist instead of loops, constructor generates these
   // Same across all SNAKokkos
@@ -203,6 +236,7 @@ public:
   Kokkos::View<int*, DeviceType> idxu_block;
   Kokkos::View<int*, DeviceType> idxu_half_block;
   Kokkos::View<int*, DeviceType> idxu_cache_block;
+  Kokkos::View<FullHalfMapper*, DeviceType> idxu_full_half;
 
 private:
   Kokkos::View<int***, DeviceType> idxz_block;
@@ -231,12 +265,12 @@ inline
   void init_rootpqarray();    // init()
 
   KOKKOS_INLINE_FUNCTION
-  void add_uarraytot(const typename Kokkos::TeamPolicy<DeviceType>::member_type& team, int, int, double, double, double, int); // compute_ui
+  void add_uarraytot(const typename Kokkos::TeamPolicy<DeviceType>::member_type& team, int, int, const real&, const real&, const real&, int); // compute_ui
 
   KOKKOS_INLINE_FUNCTION
   void compute_uarray_cpu(const typename Kokkos::TeamPolicy<DeviceType>::member_type& team, int, int,
-                      double, double, double,
-                      double, double); // compute_ui_cpu
+                      const real&, const real&, const real&,
+                      const real&, const real&); // compute_ui_cpu
 
 
   inline
@@ -246,8 +280,8 @@ inline
   int compute_ncoeff();           // SNAKokkos()
   KOKKOS_INLINE_FUNCTION
   void compute_duarray_cpu(const typename Kokkos::TeamPolicy<DeviceType>::member_type& team, int, int,
-                       double, double, double, // compute_duidrj_cpu
-                       double, double, double, double, double);
+                       const real&, const real&, const real&, // compute_duidrj_cpu
+                       const real&, const real&, const real&, const real&, const real&);
 
   // Sets the style for the switching function
   // 0 = none
@@ -259,11 +293,11 @@ inline
   int bnorm_flag;
 
   // Self-weight
-  double wself;
+  real wself;
   int wselfall_flag;
 
   int bzero_flag; // 1 if bzero subtracted from barray
-  Kokkos::View<double*, DeviceType> bzero; // array of B values for isolated atoms
+  Kokkos::View<real*, DeviceType> bzero; // array of B values for isolated atoms
 
   // for per-direction dulist calculation, specify the direction.
   int dir;