… src/KOKKOS/pair_snap_kokkos_impl.h
… src/KOKKOS/sna_kokkos_impl.h