**Summary**

This PR provides various performance optimizations to the evaluation of the SNAP potential within the Kokkos package when using the Cuda backend (though it should generalize well to any GPU), primarily featuring a large refactor of the algorithm used to compute the Wigner matrices (and derivatives thereof). This approach utilizes the AoSoA data layout introduced for some data structures in #2172, improving the consistency of the approach taken on the GPU. There are other small optimizations that improve performance of the Kokkos/SNAP-internal compressed neighbor table (which could have value elsewhere) and expose additional parallelism for small systems.  This PR does not change build options or the external interface as it stands, but there are open questions for @stanmoore1 below that may introduce changes. Further details are given under Implementation details.

This PR effectively "does no harm" to the CPU Kokkos backend (1% regressions here and there, which seems to be a theme with each GPU-specific PR). This has no effect on the CPU implementation in the SNAP package, which has been historically faster than the Kokkos/CPU backend.

Performance notes below. For robustness, I note that master is specifically commit 427d0377476a1a52570734145539b88a41cc41c9 at time of filing.

Performance improvements on Quadro V100, CUDA 11.1 for various benchmarks. Inputs are averages over 15 runs, sig figs truncated to reflect fluctuations between runs. I included measurements for different problem sizes to show benefits at smaller vs larger atom counts. Some performance equalities between `in.snap.W.2940` and `in.snap.WBe.PRB2019` are _not_ typos.

| Input file | 2Jmax | Problem Size (atoms) | Timesteps/sec master | Timesteps/sec PR | Perf improvement | Comment |
|---|---|---|---|---|---|---|
| 2J8 benchmark | 8 | 4000 | 151.911 | 271.101 | 1.78x | Benchmark problem |
| 2J14 benchmark | 14 | 4000 | 24.1521 | 26.6446 | 1.10x | ECP KPP problem |
| in.snap.W.2940 | 8 | 128 | 1580.82 | 1769.9 | 1.12x | `/examples/snap/` single atom |
| " | " | 1024 | 554.811 | 816.739 | 1.47x | increased problem size |
| in.snap.WBe.PRB2019 | 8 | 128 | 1573.58 | 1765.4 | 1.12x | `/examples/snap/` multi-atom, no chem |
| " | " | 1024 | 554.217 | 803.303 | 1.45x | increased problem size |
| in.snap.InP.JCPA2020 | 6 | 512 | 512.125 | 915.31 | 1.79x |  `/examples/snap/` multi-atom, yes chem |
| " | " | 4096 | 82.6831 | 168.69 | 2.04x |  increased problem size, batchsize 2048 |

Larger boosts for smaller 2Jmax and larger atom counts are reasonable: former due to these optimizations being targetted towards kernels scaling as `Jmax^3` as opposed to `Jmax^5`, latter due to small-system latency overheads.

**Related Issue(s)**

Further improvements on #2172. This PR (and others preceding it) establish precedence for bifurcating the CPU and GPU codepaths of the Kokkos/SNAP implementation.

**Author(s)**

Evan Weinberg (NVIDIA). eweinberg [at] nvidia.com preferred, evansweinberg [at] gmail.com for "long-lived".

**Licensing**

By submitting this pull request, I agree, that my contribution will be included in LAMMPS and redistributed under either the GNU General Public License version 2 (GPL v2) or the GNU Lesser General Public License version 2.1 (LGPL v2.1).

**Backward Compatibility**

No changes.

**Implementation Notes**

* Modifies `ComputeUi` and `ComputeFusedDeidrj` to exploit a new depth-first algorithm for computing the Wigner matrices and derivatives thereof
* Converts the data layout of `ulisttot` to AoSoA, reflecting the behavior of the new algorithm
* Eliminates `TransformYi` as it is not needed for the new data layouts used in `ComputeFusedDeidrj`
* Increases the parallelism exposed in `TransformUi`, requiring a new look-up table `idxu_full_half` which maps between symmetrized indices and full indices in the context of the symmetries of the Wigner matrices, improving performance for small systems
* Splits the Kokkos/SNAP-specific compressed neighbor table routine into separate GPU (ComputeNeigh) and CPU (ComputeNeighCPU) routines, matching previous naming conventions
* Exposes additional parallelism in the Kokkos/SNAP-specific neighbor table routine
* Splits pre-computing the Cayley-Klein parameters into a separate routine `ComputeCayleyKlein`, improving load balancing. As part of this I removed the `CayleyKleinPack` structure, replacing it with individual AoSoA structures for `a`, `b`, `da`, etc
* Updated `grow_rij` and `memory_usage` appropriately
* Replaces all internal instances of `double` with a new template parameter `real` attached to the `PairSNAPKokkos` and `SNAKokkos`
* Replaces all explicit "magic" factors of 32 (inner-most A in the AoSoA layouts, corresponding to the number of threads in a warp on an NVIDIA GPU) with a new, arguably more intuitive, template parameter `vector_length`, set to 1 (and otherwise unused) for the CPU backend

Correctness has been verified over several configurations (listed in timings above) by comparing thermodynamic quantities between Kokkos/Cuda and Kokkos/Serial runs with this PR against Kokkos/Cuda runs of the current `master` branch.

Some open points of discussion:
* I had to make a few arguably imperfect implementation choices to get the new template parameters to play nicely with the LAMMPS pair style factory and `PairStyle` declarations at the top of `src/KOKKOS/pair_snap_kokkos.h` --- for example, explicitly using `PairSNAPKokkos<LMPDeviceType, double, 32>` threw compiler errors, so I created a thin wrapper `PairSNAPKokkosDevice` class that sidesteps these issues. I'm not married to this approach, we can certainly work out something clearer or reason out the issue necessitating the wrapper.
* I didn't modify the default batch size away from `2000` as I've previously discussed with @stanmoore1 offline. I still think it'd be a good idea (at least for the GPU backend), but I wanted to revisit the details before making a blanket decision since it has implications for both the CPU and the GPU. If this gets changed it necessitates changes to the documentation, which is why I left those checkboxes below. (If anything we should change it to something that divides into a sufficiently large power of 2 --- 32 at least.)

**Post Submission Checklist**

- [ ] The feature or features in this pull request is complete
- [ ] Licensing information is complete
- [ ] Corresponding author information is complete
- [ ] The source code follows the LAMMPS formatting guidelines
- [ ] Suitable new documentation files and/or updates to the existing docs are included
- [ ] The added/updated documentation is integrated and tested with the documentation build system
- [ ] The feature has been verified to work with the conventional build system
- [ ] The feature has been verified to work with the CMake based build system

**Further Information, Files, and Links**

None.