@weinbe2 2020-12-30
A quick question @akohlmey --- I can see that the `checkstyle` test is detecting "Style issues". Is there a way that I can access the list of issues? I couldn't find them by clicking on the "Details" list (which I imagine is broadly reasonable for permissions reasons). If there's a way I can see them that I missed, or if you could please send me a copy of the list of style issues, I'd be happy to put the fixes relevant for this PR in. (It'd arguably be easier than hunting for trailing whitespaces, etc, by hand.)