@akohlmey 2020-12-30
@weinbe2 if you have difficulties accessing CI test output, you can switch from the fancy "Blue Ocean" visualization to the old style output by clicking on the `[->]` icon next to the "Login" button. That will show links with detailed outputs and - for grouped runs - links to the individual runs and from there you can access the console output.