@akohlmey 2020-12-30
@weinbe2 FYI. I just looked at the CI page myself after clicking on "Details" in the PR page. You can see the output from the whitespace checker after clicking on the yellow circle labeled "Whitespace" and then on the ">" line starting with "singularity", which has the output of the check. Some of the issues are already addressed in PR #2547, so you really only need to fix issues that would be introduced by your changes (that is why style issues are mentioned but not blocking merges (currently)).