@weinbe2 2020-12-30
@akohlmey thanks for the tip on switching back to the old style output, but also on where I could find the issues in the Blue Ocean vis style. You also headed off my questions about the MILAP formatting issues.