@stanmoore1 2021-01-04
Sure, will do.