@weinbe2 2021-01-05
@stanmoore1 what are your thoughts on increasing the default chunk size?