@weinbe2 2021-01-05
> Looks good overall, I had some minor change requests in the comments.

Thank you @stanmoore1 ! I've addressed them up to a few open questions. I also did a quick compile-and-correctness check.