@weinbe2 2021-01-05
> > what are your thoughts on increasing the default chunk size?
> 
> I think we should increase it. Is 4000 a good value?

Let's use 4096, the motivation being problem sizes with larger powers of two. (For example---with a block size of 4000, a problem size of 4096 will have one full chunk then one latency-driven chunk with 96; 8192 will have two full chunks then one latency-driven of size 192, etc.)

If you're okay with that, I can make the change across the `SNAP` package, documentation, I have the PR you introduced the chunk size in for reference.