@stanmoore1 2021-01-05
4096 sounds good. I can make the changes, I'm working on this now.