@stanmoore1 2021-01-05
I'm double-checking performance and correctness on Summit now.