@weinbe2 2021-01-05
> I'm double-checking performance and correctness on Summit now.

Thanks for making the changes @stanmoore1 , I'll do a quick local check too.