@stanmoore1 2021-01-05
On Summit I see a 1.6x speedup for 2J8 and 1.2x speedup for 2J14. This is for 2000 atoms instead of 4000 atoms though.