@weinbe2 2021-01-05
> On Summit I see a 1.6x speedup for 2J8 and 1.2x speedup for 2J14. This is for 2000 atoms instead of 4000 atoms though.

That all seems reasonable---slightly different GPUs and CUDA versions, different arch w/increased latencies, different atom count, and replacing the `ComputeUi` and `ComputeFusedDeidrj` team size with a `constexpr` gave a slight speed-up not reflected in my numbers on top.