@stanmoore1 2021-01-05
Yeah, I'm including the ~5% speedup from CUDA version too.