@stanmoore1 2021-01-05
I'm seeing slight differences in the forces/trajectories between GPU and vanilla CPU. It shows up in latest master too, so this PR isn't the source, but we aren't getting the agreement I would expect for only 1 timestep. This seems like a small numerical regression, any ideas?