@stanmoore1 2021-01-05
I turned off fix Langevin for the correctness check, so the pRNG isn't the source.