@athomps 2021-01-05
There were issues with velocity assignment earlier.  Are you using 
```velocity all create 300.0 4928459 loop geom```
?