@athomps 2021-01-05
`loop geom` eliminates effect of MPI ranks.