@stanmoore1 2021-01-05
I verified correctness. I have one or two more checks, then should be ready to merge.