@weinbe2 2021-01-05
> Yeah, I'm including the ~5% speedup from CUDA version too.

Using CUDA 11.2 and 2000 atoms, I'm seeing (in units timestep/sec)

|Problem size | master | PR | speed-up |
|---|---|---|---|
|2J8|350|548|1.56x|
|2J14|52|60|1.15x|

And for 4000 atoms

|Problem size | master | PR | speed-up |
|---|---|---|---|
|2J8|154|300|1.95x|
|2J14|26|30|1.15x|

* 2000 atoms just isn't saturating the GPU for the 2J8 problem
* I know the 11.2 release lead to some non-trivial speed-ups for the AoSoA-style kernels, so I'm not surprised re-writing `ComputeUi` and `ComputeFusedDeidrj` in that form gave an impressive boost relative to 11.1.