@weinbe2 2021-01-05
> I get 525 and 66 timesteps/s on Summit. Seems weird that 2J14 is better on Power9?

I'm using a Quadro instead of a Tesla, so it's reasonable.