@stanmoore1 2021-01-05
I believe this is ready for merge. Thanks @weinbe2 for the impressive speedups!